Module chatbot::handler

Contains the MessageHandler trait and handler implementations.

Modules

sup

Structs

BasicResponseHandler
A basic response handler

GithubIssueLinker
Respond to github issue links with the title and hyperlink

Traits

MessageHandler
Implementing a MessageHandler enables responding to IncomingMessages. There are currently very few requirements to creating a handler. The handle function receives IncomingMessages. The implementer is responsible for testing whether it's interested in replying. Optionally call reply on the IncomingMessage to send a response.

Type Definitions

HandlerResult