Struct chatbot::handler::sup::AddMatch

pub struct AddMatch {
    // some fields omitted
}

Add startuppong.com matches from chat

Say "add a match where Player A beat Player B" and a match will be recorded on startuppong.com. The handler should reply with the old and new ranks for each player.

Methods

impl AddMatch

fn new(account: Account) -> AddMatch