Struct chatbot::message::IncomingMessage

pub struct IncomingMessage {
    // some fields omitted
}

adapters convert strings they receive into an IncomingMessage. The properties on this struct exist to help adapters route any OutgoingMessage back to where the IncomingMessage originated.

Types implementing the MessageHandler trait should use IncomingMessage.reply to respond. It is up to the ChatAdapter to decide what reply means in the context of the service it provides.

Methods

impl IncomingMessage

fn new(from_adapter: String, server: Option<String>, channel: Option<String>, user: Option<String>, message: String, sender: Sender<AdapterMsg>) -> IncomingMessage

fn channel(&self) -> Option<&str>

fn user(&self) -> Option<&str>

fn get_contents(&self) -> &str

fn reply(&self, msg: String) -> Result<(), SendError<AdapterMsg>>

Reply to the message.