Struct chatbot::Chatbot

pub struct Chatbot {
    // some fields omitted
}

The Chatbot is the central data structure of the chatbot platform. It contains a run method which listens for messages from adapters and routes them to handlers. Any program which uses chatbot will need to minimally create a Chatbot, add an adapter, add a handler, and call Chatbot run.

Methods

impl Chatbot

fn new(name: &str) -> Chatbot

Create a new chatbot instance.

The name provided here will be used for message filtering and in handlers for whatever they want. You'll want to make your binding mutable so you can call add_adapter and add_handler.

fn get_name(&self) -> &str

Return the name provided in initialization

fn get_addresser(&self) -> &Regex

Return the regular expression of what the bot will be addressed by

fn add_adapter<T>(&mut self, adapter: T) where T: ChatAdapter + 'static

Add a ChatAdapter to the bot

Add as many adapters as you like. The IncomingMessages sent by adapters are made available to all handlers regardless of how many adapters exist. The IncomingMessage.reply method makes sure the response is sent back to the adapter from whence the message came.

fn add_handler<T>(&mut self, handler: T) where T: MessageHandler + 'static

Add a MessageHandler to the bot

The more handlers you have the more useful your bot becomes (for potentially loose definitions of useful :P). Check out the handler! macro for making simple handlers and also see the MessageHandler implementors section for a list of built in handlers.

fn add_addressed_handler<T>(&mut self, handler: T) where T: MessageHandler + 'static

Add a MessageHandler, that requires the bot to be addressed, to the bot

fn run(&self)

Start processing messages

Call process_events on all of the adapters and recv on the IncomingMessage channel. Distribute IncomingMessages to list of handlers.