Trait chatbot::handler::MessageHandler

pub trait MessageHandler {
    fn name(&self) -> &str;
    fn handle(&self, incoming: &IncomingMessage) -> HandlerResult;
    fn re(&self) -> &Regex;

    fn can_handle(&self, msg: &str) -> bool { ... }
    fn get_captures<'a>(&self, msg: &'a str) -> Option<Captures<'a>> { ... }
}

Implementing a MessageHandler enables responding to IncomingMessages. There are currently very few requirements to creating a handler. The handle function receives IncomingMessages. The implementer is responsible for testing whether it's interested in replying. Optionally call reply on the IncomingMessage to send a response.

Example

A simple echo handler might look something like the following:

use chatbot::handler::MessageHandler;
use chatbot::handler::HandlerResult;
use chatbot::message::IncomingMessage;

use regex::Regex;

struct EchoHandler {
    regex: Regex
}

impl EchoHandler {
    fn new() -> EchoHandler {
        EchoHandler {
            regex: Regex::new(r".").unwrap()
        }
    }
}

impl MessageHandler for EchoHandler {
    fn name(&self) -> &str {
        "echo"
    }

    fn re(&self) -> &Regex {
        &self.regex
    }

    fn handle(&self, incoming: &IncomingMessage) -> HandlerResult {
        let response = incoming.get_contents().to_owned();
        Ok(try!(incoming.reply(response)))
    }
}

Required Methods

fn name(&self) -> &str

fn handle(&self, incoming: &IncomingMessage) -> HandlerResult

fn re(&self) -> &Regex

Provided Methods

fn can_handle(&self, msg: &str) -> bool

Uses re() to test whether the handler should process this message.

fn get_captures<'a>(&self, msg: &'a str) -> Option<Captures<'a>>

Uses re() to get capturing groups from a message